HealBot updated!! 27-05-2022 Version 1.1
Fixed:
The buff tracker not showing all buffs.

Issues:
Dispel Magic (mage spell) not showing up in the list of debuff spells, this seems to be a issue with the original version wich the devs fixed in later version which offcouse dont work on 3.3.5. (Going to look in to this).